// Models/Enums.ts
export enum SaleStatus {
    PENDING = "pending",
    APPROVED = "approved",
    DECLINED = "declined",
    CANCELLED = "cancelled"
  }
  
  export enum PayoutStatus {
    PENDING = "pending",
    COMPLETED = "completed",
    CANCELLED = "cancelled",
    REJECTED = "rejected",
    FAILED = "failed"
  }

  export enum WithdrawStatus {
    PENDING = "pending",
    COMPLETED = "completed",
    CANCELLED = "cancelled",
    REJECTED = "rejected"
  }
  